support to leave a specific instance

+++ bin/leave-testnet.sh
@@ -26,9 +26,15 @@ PROJECT_PATH=`pwd`/
 
 DOMAIN=${DOMAIN:-testnet.diva.i2p}
 
+LEAVE_INSTANCE=${LEAVE_INSTANCE:0}
+
 for nameFile in `ls -1 ${PROJECT_PATH}data/ | fgrep ".${DOMAIN}"`
 do
   INSTANCE=$(<${PROJECT_PATH}data/${nameFile})
+  if [[ ${LEAVE_INSTANCE} -gt 0 && ! ${LEAVE_INSTANCE} -eq ${INSTANCE} ]]
+  then
+    continue
+  fi
   rm -f ${PROJECT_PATH}data/${nameFile}
   IDENT=nx${INSTANCE}
